fix(patterns): carry a 128-bit literal in pattern position (B-2026-08-20-4)
`LiteralPattern::Integer` held an `i64` while `ExprKind::Integer` holds an
`i128`, so the two positions disagreed about what an integer literal is. Every
value past `i64::MAX` — the upper half of `i128`, the whole top half of `u128`,
and any `u128` beyond `i128::MAX` — could be written as an expression and not as
a pattern. A 128-bit scrutinee could be compared and bound but never matched
against its own literal, and the only workaround was a guard (`n if n == …`),
which exhaustiveness cannot see through.

Widening the node fixes it, and the widening is almost all of the change: five
compile errors, every one of them in `parser/patterns.rs`. Every other consumer
was already payload-generic or already working in `i128` — including
`exhaustive.rs`, which the row expected to need the most care. Its range
arithmetic runs in `i128` space already (`PatCtor::IntRange`), and `int_domain`
deliberately returns `None` for the 128-bit widths, so those route through the
open-domain path that demands an explicit wildcard: imprecise but sound, and
unchanged by this. Measured rather than assumed — a match on `i128` with no
wildcard is still correctly rejected, and the `u8`/`u64` domains still enumerate.

The parser's two refusals are gone and its out-of-range carrier helper now
covers all four unsigned bands plus both 128-bit suffixes. `i128` takes the
magnitude positively; `u128` wraps its top half into the signed carrier, which
is exactly what `parser/exprs.rs` does for the expression spelling — the two
encodings have to agree or a pattern could never match its own value.

Clippy's `--all-targets` and `--features llvm` legs then found eight now-identity
widening conversions across five files, which are removed: they are the sites
that used to widen the payload by hand. Two of the eight are visible only on the
llvm leg.

Verified byte-identical across `karac build`, that build with KARAC_AUTO_PAR=0,
`karac run` (JIT) and `karac run --interp`: literal arms at `i128::MAX` and 2^100,
a 128-bit range pattern, `u128::MAX`, and `170141183460469231731687303715884105728u128`
— one past `i128::MAX`, so its carrier value is negative — distinguished from
both `u128::MAX` and `5u128` in the same match. The MISS arms are asserted
alongside the hits, since a wrapped encoding that disagreed between pattern and
scrutinee would show up as a pattern that fires when it should not.

tests/parser.rs's `a_128bit_literal_pattern_is_a_diagnostic_not_a_compiler_panic`
asserted the refusal this removes; it is now
`a_128bit_literal_pattern_parses` and covers the range-bound path too.

Two gaps found while measuring, both pre-existing and both filed separately: a
NEGATIVE literal pattern does not parse at any width (`match n { -5 => … }`),
and a `u64` range pair that covers the whole domain is wrongly reported
non-exhaustive because the top-half bounds ride the carrier as negatives.

+++ src/ast/patterns.rs
@@ -237,7 +237,17 @@ pub enum RestPattern {
 
 #[derive(Debug, Clone)]
 pub enum LiteralPattern {
-    Integer(i64, Option<IntSuffix>),
+    /// Payload is `i128` for the same reason `ExprKind::Integer`'s is: a
+    /// 128-bit literal has to be representable in the node before any consumer
+    /// can act on it. While it was `i64` no literal past `i64::MAX` could be a
+    /// pattern at all — the whole upper half of `i128`/`u128`, plus the entire
+    /// `u64` top half, were reachable in expression position and unreachable in
+    /// pattern position (B-2026-08-20-4).
+    ///
+    /// An UNSIGNED value keeps riding as its wrapped two's-complement pattern,
+    /// exactly as the expression carrier does, so a pattern and its scrutinee
+    /// compare as identical bits without either side re-reading the sign.
+    Integer(i128, Option<IntSuffix>),
     Float(f64, Option<FloatSuffix>),
     Char(char),
     String(String),
